LayoutTests/media/video-seeking-expected.txt

Test that seeking attribute is true in timeupdate event after seek and goes back to false when seeking completes

EVENT(load)
TEST(!video.seeking) OK
RUN(video.currentTime = 2)
EVENT(timeupdate)
TEST(video.seeking) OK
TEST(video.currentTime == 2) OK
TEST(!video.seeking) OK
END OF TEST
